% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_warrior_data_file_names.R
\name{get_latest_warrior_data_file_names}
\alias{get_latest_warrior_data_file_names}
\title{Gets the latest downloaded files (in terms of date) from the provided
from_directory, as well as the date reflected in the filenames.
If no arguments are provided, the directory is assumed to be the '~/Downloads/' folder.}
\usage{

  get_latest_warrior_data_file_names(from_directory = "~/Downloads/warrior_data/")
}
\arguments{
\item{from_directory}{A directory where to find the latest data.}
}
\value{
A list containing of key-value pairs
        returned_list[['data']] to get the data file name,
        returned_list[['metadata']] to get the metadata file name and,
        returned_list[['date']] to get the date associated with the data.
}
\description{
Gets the latest downloaded files (in terms of date) from the provided
from_directory, as well as the date reflected in the filenames.
If no arguments are provided, the directory is assumed to be the '~/Downloads/' folder.
}
\examples{
get_latest_warrior_data_file_names()
}
